style: move constants onto shared folder
